FIXED: all Bugs reported reported by Shaun

+++ src/Utilities/deletePromocode.js
@@ -2,10 +2,12 @@ import db from "../firebase"
 import setMessage from "./setMessage"
 
 const deletePromocode = code => {
-  db.doc(`promocodes/${code}`)
-    .delete()
-    .then(() => setMessage('Promocode Deleted!'))
-    .catch(err => setMessage(err?.message || 'Unable to Delete!'));
+  if(window.confirm('Delete Promocode?')){
+    db.doc(`promocodes/${code}`)
+      .delete()
+      .then(() => setMessage('Promocode Deleted!'))
+      .catch(err => setMessage(err?.message || 'Unable to Delete!')); 
+  }
 }
 
 export default deletePromocode;
